requirements.txt
art==5.8
colorama==0.4.6
fake-useragent==1.1.1
requests==2.28.1
speedtest-cli==2.1.3
termcolor==2.1.1



